Feat: seminar7 exercise

+++ post/views.py
@@ -1,10 +1,11 @@
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from django.shortcuts import render
-from .models import Post
+from .models import Post, Like
 from rest_framework.views import APIView
 from rest_framework import status
 from .serializers import PostSerializer
+from tag.models import Tag
 
 # Create your views here.
 
@@ -30,6 +31,8 @@ def CreatePostView(request):
     return Response({"msg": f"'{post.title}'이 생성되었어요!"})
 
 # 아래에 추가해주세요!
+
+
 @api_view(['GET'])
 def ReadAllPostView(request):
     posts = Post.objects.all()
@@ -54,11 +57,24 @@ class PostListView(APIView):
         # return Response(contents, status=status.HTTP_200_OK)
 
     def post(self, request):
+        author = request.user
         title = request.data.get('title')
         content = request.data.get('content')
+        tag_ids = request.data.get('tags')
+
+        if not author.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
         if not title or not content:
             return Response({"detail": "[title, content] fields missing."}, status=status.HTTP_400_BAD_REQUEST)
-        post = Post.objects.create(title=title, content=content)
+
+        for tag_id in tag_ids:
+            if not Tag.objects.filter(id=tag_id).exists():
+                return Response({"detail": "Provided tag not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        post = Post.objects.create(title=title, content=content, author=author)
+
+        post.tags.set(tag_ids)
 
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -92,6 +108,9 @@ class PostDetailView(APIView):
             post = Post.objects.get(id=post_id)
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -108,3 +127,42 @@ class PostDetailView(APIView):
             return Response(serializer.data)
         else:
             return Response({"msg": "title or content is missing!"})
+
+    def patch(self, request, post_id):
+        try:
+            post = Post.objects.get(id=post_id)
+        except:
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        serializer = PostSerializer(post, data=request.data, partial=True)
+        if not serializer.is_valid():
+            return Response({"detail": "data validation error"}, status=status.HTTP_400_BAD_REQUEST)
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+
+class LikeView(APIView):
+    def post(self, request, post_id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        ### 1 ###
+        try:
+            post = Post.objects.get(id=post_id)
+        except:
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        ### 2 ###
+        like_list = post.like_set.filter(user=request.user)
+
+        ### 3 ###
+        if like_list.count() > 0:
+            post.like_set.get(user=request.user).delete()
+        else:
+            Like.objects.create(user=request.user, post=post)
+
+        serializer = PostSerializer(instance=post)
+        return Response(serializer.data, status=status.HTTP_200_OK)
